+++ point-of-sale-sdk/src/test/java/com/squareup/sdk/pos/RealPosClientTest.java
@@ -43,7 +43,7 @@ import static com.squareup.sdk.pos.PosApi.EXTRA_CUSTOMER_ID;
 import static com.squareup.sdk.pos.PosApi.EXTRA_LOCATION_ID;
 import static com.squareup.sdk.pos.PosApi.EXTRA_NOTE;
 import static com.squareup.sdk.pos.PosApi.EXTRA_POINT_OF_SALE_CLIENT_ID;
-import static com.squareup.sdk.pos.PosApi.EXTRA_REQUEST_METADATA;
+import static com.squareup.sdk.pos.PosApi.EXTRA_REQUEST_STATE;
 import static com.squareup.sdk.pos.PosApi.EXTRA_SDK_VERSION;
 import static com.squareup.sdk.pos.PosApi.EXTRA_TENDER_CARD;
 import static com.squareup.sdk.pos.PosApi.EXTRA_TENDER_TYPES;
@@ -153,7 +153,7 @@ public class RealPosClientTest {
         .autoReturn(true)
         .enforceBusinessLocation("location")
         .customerId("customerId")
-        .requestMetadata("metadata")
+        .state("state")
         .note("note")
         .build();
 
@@ -164,7 +164,7 @@ public class RealPosClientTest {
     assertThat(intent.getStringExtra(EXTRA_POINT_OF_SALE_CLIENT_ID)).isEqualTo(CLIENT_ID);
     assertThat(intent.getStringExtra(EXTRA_NOTE)).isEqualTo("note");
     assertThat(intent.getStringExtra(EXTRA_API_VERSION)).isEqualTo("v3.0");
-    assertThat(intent.getStringExtra(EXTRA_REQUEST_METADATA)).isEqualTo("metadata");
+    assertThat(intent.getStringExtra(EXTRA_REQUEST_STATE)).isEqualTo("state");
     assertThat(intent.getStringExtra(EXTRA_LOCATION_ID)).isEqualTo("location");
     assertThat(intent.getStringExtra(EXTRA_CUSTOMER_ID)).isEqualTo("customerId");
     assertThat(intent.getStringArrayListExtra(EXTRA_TENDER_TYPES)).containsExactly(
@@ -183,7 +183,7 @@ public class RealPosClientTest {
     assertThat(intent.hasExtra(EXTRA_API_VERSION)).isTrue();
     assertThat(intent.hasExtra(EXTRA_SDK_VERSION)).isTrue();
     assertThat(intent.hasExtra(EXTRA_CURRENCY_CODE)).isTrue();
-    assertThat(intent.hasExtra(EXTRA_REQUEST_METADATA)).isTrue();
+    assertThat(intent.hasExtra(EXTRA_REQUEST_STATE)).isTrue();
     assertThat(intent.hasExtra(EXTRA_TENDER_TYPES)).isTrue();
 
     assertThat(intent.hasExtra(EXTRA_CUSTOMER_ID)).isFalse();